agregar el role premium

+++ src_Capas/Controles/controllers/ProductsControllers.js
@@ -1,3 +1,4 @@
+import { logDebug, logError, logInfo } from '../../Errores/Winston.js';
 import { createError, errors } from '../../Errores/errorModule.js';
 import ProductRepository from '../../Persistencia/DAO/ProductRepository.js';
 
@@ -17,18 +18,34 @@ export const getProducts = async (req, res) => {
 export const saveProduct = async (req, res) => {
     try {
         const { name, photo, price, category, description } = req.body;
-        const newProduct = await productRepository.createProduct(name, photo, price, category, description);
-        res.status(200).json({
-            status: 'success',
-            message: 'Producto guardado exitosamente.',
-            product: newProduct,
-        });
+        // Obtener el _id del usuario desde req.user._id
+        const ownerId = req.user._id;
+        
+        // Verificar si el usuario es premium
+        if (req.user.role === 'premium') {
+            // Crear el producto con el owner del usuario premium
+            const newProduct = await productRepository.createProduct(name, photo, price, category, description, ownerId);
+            res.status(200).json({
+                status: 'success',
+                message: 'Producto guardado exitosamente.',
+                product: newProduct,
+            });
+        } else {
+            // Si el usuario no es premium, guardar el producto sin owner
+            const newProduct = await productRepository.createProduct(name, photo, price, category, description);
+            res.status(200).json({
+                status: 'success',
+                message: 'Producto guardado exitosamente.',
+                product: newProduct,
+            });
+        }
     } catch (error) {
         console.error(error);
         const errorDetails = createError(errors.PRODUCT_SAVE_ERROR);
         res.status(errorDetails.status).json({ status: 'error', error: errorDetails.message });
     }
 };
+
 
 export const getProductByIdController = async (req, res) => {
     try {
@@ -45,14 +62,22 @@ export const getProductByIdController = async (req, res) => {
     }
 };
 
+// Controlador para eliminar un producto por ID
 export const deleteProductByIdController = async (req, res) => {
     try {
         const productId = req.params.id;
-        await productRepository.deleteProductById(productId);
-        res.status(200).json({
-            status: 'success',
-            message: 'Producto eliminado exitosamente.',
-        });
+        const product = await productRepository.getProductById(productId);
+        
+        // Verificar si el usuario es administrador o premium y si es dueño del producto
+        if (req.user.role === 'admin' || (req.user.role === 'premium' && product.owner.equals(req.user._id))) {
+            await productRepository.deleteProductById(productId);
+            res.status(200).json({
+                status: 'success',
+                message: 'Producto eliminado exitosamente.',
+            });
+        } else {
+            return res.status(403).json({ status: 'error', message: 'Acceso no autorizado para eliminar este producto.' });
+        }
     } catch (error) {
         console.error(error);
         const errorDetails = createError(errors.PRODUCT_DELETE_ERROR);
@@ -60,13 +85,29 @@ export const deleteProductByIdController = async (req, res) => {
     }
 };
 
-export const updateProductByIdController = async (productId, updatedProductData) => {
+// Controlador para actualizar un producto por ID
+export const updateProductByIdController = async (req, res) => {
+    const productId = req.params.id;
+    const updatedProductData = req.body;
+
     try {
-        const updatedProduct = await productRepository.updateProductById(productId, updatedProductData);
-        return updatedProduct;
+        const product = await productRepository.getProductById(productId);
+
+        // Verificar si el usuario es administrador o premium y si es dueño del producto
+        if (req.user.role === 'admin' || (req.user.role === 'premium' && product.owner.equals(req.user._id))) {
+            const updatedProduct = await productRepository.updateProductById(productId, updatedProductData);
+            res.status(200).json({
+                status: 'success',
+                message: 'Producto actualizado exitosamente.',
+                product: updatedProduct,
+            });
+        } else {
+            return res.status(403).json({ status: 'error', message: 'Acceso no autorizado para actualizar este producto.' });
+        }
     } catch (error) {
         console.error(error);
         const errorDetails = createError(errors.PRODUCT_UPDATE_ERROR);
-        throw errorDetails;
+        res.status(errorDetails.status).json({ status: 'error', error: errorDetails.message });
     }
 };
+
